refactor: path를 Station이 갖는 구조에서 PathRepository가 따로 관리하도록 수정
