+++ src/main/java/subway/domain/line/Line.java
@@ -1,12 +1,6 @@
 package subway.domain.line;
 
-import subway.domain.station.Station;
-
-import java.util.ArrayList;
-import java.util.List;
-
 public class Line {
-    private List<Station> stations = new ArrayList<>();
     private String name;
 
     public Line(String name) {
