+++ src/main/java/subway/domain/path/PathService.java
@@ -1,0 +1,10 @@
+package subway.domain.path;
+
+public class PathService {
+    private PathService() {
+    }
+
+    public static void addPath(Path path) {
+
+    }
+}
