+++ src/main/java/subway/domain/station/Station.java
@@ -1,11 +1,7 @@
 package subway.domain.station;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class Station {
     private String name;
-    private List<Path> paths = new ArrayList<>();
 
     public Station(String name) {
         this.name = name;
@@ -13,9 +9,5 @@ public class Station {
 
     public String getName() {
         return name;
-    }
-
-    public void addPath(Path path) {
-        paths.add(path);
     }
 }
